mtd: nand: sunxi_nand.c: Use common struct mtd pointer
The common nand infrastructure expects a common mtd pointer, allocated
in nand.c. This patch now uses this pointer and does not allocate
a new mtd structure in this driver as well.

Also this assignment is moved before the sunxi_nand_chip_set_timings()
call, since this call into the common nand interface and its structures
need to be valid by then.

Signed-off-by: Stefan Roese <sr@denx.de>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>